add ability to add family attendance
