+++ app/schemas/attendance.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class AttendanceData(BaseModel):
+    id: str
+    hours: int
+
+
+class SetAttendanceRequest(BaseModel):
+    attendance: list[AttendanceData]
